requirements.txt
Django==1.11.3
PyGithub==1.35
PyJWT==1.5.2
pytz==2017.2
